fix(lifecycle): reconcile merge — adapt to Option-returning expand_targets, dedup tests, split run command out of commands.rs to satisfy line limit
